SangJin 통합
