--- web/src/sangjin/components/MainPage/AttendanceCheck/AttendanceCheck.css
@@ -1,133 +1,0 @@
-.card_attchk {
-	display: flex;
-	flex-direction: column;
-	border-radius: 8px;
-	margin-top: 17px;
-	font-size: 12px;
-	background-color: #eaeaea;
-}
-
-.card_attchk_head {
-	display: flex;
-	flex-wrap: wrap;
-	justify-content: space-between;
-}
-
-.card_attchk_head_title {
-	margin-top: 16px;
-	margin-left: 16px;
-	margin-bottom: 14px;
-	font-family: "Inter";
-	font-style: normal;
-	font-weight: 700;
-	font-size: 16px;
-	line-height: 24px;
-	text-align: left;
-	letter-spacing: 0.44px;
-	color: #000000;
-}
-
-.card_attchk_onul {
-	margin-top: 17px;
-	margin-right: 9px;
-	width: 124px;
-	height: 22px;
-	background: #d9d9d9;
-	border-radius: 5px;
-	line-height: 22px;
-	font-family: "Inter";
-	font-style: normal;
-	font-weight: 700;
-	font-size: 12px;
-	line-height: 22px;
-	letter-spacing: -0.408px;
-	color: #000000;
-	display: flex;
-	flex-wrap: wrap;
-	justify-content: center;
-}
-.attchk_coin {
-	margin-left: 6px;
-	margin-right: 2px;
-	margin-top: 2px;
-	background-color: transparent;
-	border-width: 0px;
-	width: 16px;
-	height: 16px;
-}
-.card_attendance {
-	margin-left: 6px;
-	margin-right: 7px;
-	margin-bottom: 8px;
-	border-radius: 9px;
-	height: 40px;
-	background-color: #7676803d;
-}
-/* .card_attchk_table {
-	position: absolute;
-	margin-left: 1.749%;
-	margin-right: 2.04%;
-	margin-bottom: 8px;
-	padding-left: 8px;
-	padding-top: 4.5px;
-	padding-bottom: 4.5px;
-	padding-right: 8px;
-	border-radius: 9px;
-	height: 40px;
-	vertical-align: middle;
-	background-color: #7676803d;
-	border-spacing: 0px;
-	width: 96.2%;
-}
-.card_attchk_table td {
-	border-right: 1px rgba(84, 84, 88, 0.65) solid;
-}
-.card_attchk_table td:first-child {
-  border-left: 0;
-}
-.card_attchk_table td:last-child {
-	border-right: 1px transparent solid;
-} */
-
-.attchek_first {
-	height: 31px;
-	background-color: #636366;
-	border-top-left-radius: 9px;
-	border-bottom-left-radius: 9px;
-}
-.attchek_middle {
-	height: 31px;
-	background-color: #636366;
-	/* border-radius: 9px; */
-}
-.attchek_last {
-	height: 31px;
-	background-color: #636366;
-	border-top-right-radius: 9px;
-	border-bottom-right-radius: 9px;
-}
-.attchek_yes {
-	height: 31px;
-	background-color: #636366;
-	border-radius: 9px;
-}
-/* .tablediv {
-	position: relative;
-} */
-.card_attchk_tablein {
-	/* position: relative; */
-	margin-left: 1.749%;
-	margin-right: 2.04%;
-	margin-bottom: 8px;
-	padding-left: 8px;
-	padding-top: 4.5px;
-	padding-bottom: 4.5px;
-	padding-right: 8px;
-	border-radius: 9px;
-	height: 40px;
-	vertical-align: middle;
-	background-color: #7676803d;
-	color: white;
-	border-spacing: 0px;
-	width: 96.2%;
-}
