+++ web/src/sangjin/components/MainPage/MainPage.css
@@ -1,24 +1,19 @@
 .header {
-  font-family: "Inter";
-  font-style: normal;
-  font-weight: 700;
-  font-size: 34px;
-  line-height: 41px;
-  /* identical to box height, or 121% */
-
-  letter-spacing: 0.374px;
-
-  color: #000000;
-  text-align: left;
-  margin-left: 30px;
-  margin-top: 42px;
-  margin-bottom: 22px;
-  /* identical to box height, or 121% */
-
-  letter-spacing: 0.374px;
-  color: #000000;
+	font-family: "Inter";
+	font-style: normal;
+	font-weight: 700;
+	font-size: 34px;
+	line-height: 41px;
+	letter-spacing: 0.374px;
+	color: #000000;
+	text-align: left;
+	margin-left: 30px;
+	margin-top: 42px;
+	margin-bottom: 22px;
+	letter-spacing: 0.374px;
+	color: #000000;
 }
 .contents {
-  margin-left: 23px;
-  margin-right: 23px;
+	margin-left: 23px;
+	margin-right: 23px;
 }
