+++ web/src/sangjin/components/MainPage/MainPage.tsx
@@ -1,39 +1,39 @@
-import { platform } from "os";
 import React, { useCallback, useEffect, useState } from "react";
 import "./MainPage.css";
 import Newsletter from "./Newsletter/Newsletter";
-import { getNewsletterList } from "./Newsletter/NewsletterList";
-import { Trend } from "./Trend";
 import ChnageMBTI from "./ChangeMBTI";
-import HashTag from "./HashtagSearchResult";
-import { getRandomList, RandomListInit } from "./RandomList";
+import { RandomListInit } from "./RandomList";
 import HashtagSearchResult from "./HashtagSearchResult";
-import MyBettingRecord from "./MyBettingRecord/MyBettingRecord";
-import AttendanceCheck from "./AttendanceCheck/AttendanceCheck";
-import TopHashtag from "./TopHashTag/TopHashtag";
-export function MainPage() {
-	let Newsletter_example: any[] = [];
-	let RandomList: any[] = [];
-	RandomList = getRandomList();
 
+export function MainPage() {
+	let RandomList: any[] = [];
 	let [result, setResult] = useState<any[]>([]);
 	let [item, setItem] = useState<any[]>([]);
 	let [isLoading, setIsLoading] = useState<boolean>(true);
 
+	const getFetchData = async () => {
+		// randomList에서 화면에 띄울 데이터(Result)에 넣고 나머지는 Item에 저장
+		setResult([...RandomList.slice(0, 7)]);
+		setItem([...RandomList.slice(7)]);
+		setIsLoading(false);
+	};
+
 	const fetchMoreData = async () => {
+		//스크롤 다운했을때 Item에서 7개씩 빼서 Result에 추가
 		setIsLoading(true);
-		let tmp = item.slice(0, 5);
-		for (let i = 0; i < 5; i++) {
+		let tmp = item.slice(0, 7);
+		for (let i = 0; i < 7; i++) {
 			if (tmp[i] != undefined) {
 				result.push(tmp[i]);
 			}
 		}
 		setResult([...result]);
-		item = item.slice(5);
+		item = item.slice(7);
 		setIsLoading(false);
 	};
 
 	const _infiniteScroll = useCallback(() => {
+		//스크롤 이벤트
 		let scrollHeight = Math.max(
 			document.documentElement.scrollHeight,
 			document.body.scrollHeight
@@ -48,18 +48,16 @@ export function MainPage() {
 			fetchMoreData();
 		}
 	}, [isLoading]);
-	const getFetchData = async () => {
-		setResult([...RandomList.slice(0, 7)]);
-		setItem([...RandomList.slice(7)]);
-		setIsLoading(false);
-	};
 
 	useEffect(() => {
+		//스크롤 이벤트 등록
 		window.addEventListener("scroll", _infiniteScroll, true);
 		return () => window.removeEventListener("scroll", _infiniteScroll, true);
 	}, [_infiniteScroll]);
+
 	useEffect(() => {
-		RandomListInit();
+		//페이지 랜더링시 실행될 function
+		RandomList = RandomListInit();
 		getFetchData();
 	}, []);
 
@@ -68,21 +66,14 @@ export function MainPage() {
 			<div className="header">뉴스피드</div>
 			<div className="contents">
 				{result.map(function (i): JSX.Element {
-					console.log("result map i .............", i);
-					if (i.type === "attendancecheck") {
-						return <AttendanceCheck />;
-					} else if (i.type === "changembti") {
+					//result에 있는 컴포넌트 mainpage에 띄우기
+					if (i.type === "changembti") {
 						return <ChnageMBTI />;
-					} else if (i.type === "mybettingrecord") {
-						return <MyBettingRecord />;
 					} else if (i.type === "hashtagsearchresult") {
 						return <HashtagSearchResult />;
-					} else if (i.type === "tophashtag") {
-						return <TopHashtag />;
 					} else {
 						return (
 							<Newsletter
-								// key={i.data.newsletterId}
 								title={i.data.title}
 								thumbnail={i.data.thumbnail}
 								url={i.data.url}
