--- web/src/sangjin/components/MainPage/MyBettingRecord/MyBettingRecord.css
@@ -1,33 +1,0 @@
-.card_mybat {
-  margin-top: 17px;
-  display: flex;
-  flex-direction: column;
-  border-radius: 8px;
-  background-color: #eaeaea;
-}
-.card_mybat_header {
-  margin-top: 16px;
-  margin-left: 16px;
-  margin-bottom: 12px;
-  font-family: "Inter";
-  font-style: normal;
-  font-weight: 700;
-  font-size: 16px;
-  line-height: 24px;
-  display: flex;
-  align-items: center;
-  letter-spacing: 0.44px;
-  color: #000000;
-}
-.card_mybat_contents {
-  margin-left: 23px;
-  margin-right: 23px;
-  text-align: left;
-  font-family: "Inter";
-  font-style: normal;
-  font-weight: 400;
-  font-size: 14px;
-  line-height: 22px;
-  letter-spacing: -0.408px;
-  color: #434343;
-}
