+++ web/src/sangjin/components/MainPage/Newsletter/NewsletterType.tsx
@@ -1,0 +1,8 @@
+export type NewsletterType = {
+	newsletterId: string;
+	thumbnail: string;
+	title: string;
+	hashtag: { hashtag_id: string; name: string }[];
+	url: string;
+	writer: string;
+};
