--- web/src/sangjin/components/MainPage/TopHashTag/TopHashtag.css
@@ -1,54 +1,0 @@
-.tophashtag_card_item {
-  display: flex;
-  flex-direction: column;
-  border-radius: 8px;
-  margin-top: 17px;
-  font-size: 12px;
-  background-color: #eaeaea;
-  /* width: 343px;
-    height: 315px;
-    background-color: #eaeaea;
-    overflow: auto;
-    margin-top: 17px; */
-}
-.tophashtag_card_header {
-  margin-top: 16px;
-  margin-left: 16px;
-  margin-bottom: 14px;
-  /* font-family:  */
-  font-family: "Inter";
-  font-style: normal;
-  font-weight: 700;
-  font-size: 16px;
-  line-height: 24px;
-  text-align: left;
-  color: #000000;
-  letter-spacing: 0.44px;
-}
-
-.tophashtag_card_tag {
-  margin-bottom: 11px;
-  display: flex;
-  justify-content: space-around;
-  margin-left: 25px;
-  margin-right: 23px;
-  text-align: left;
-  color: #434343;
-}
-
-.tophashtag_hashtag {
-  font-family: "Inter";
-  font-style: normal;
-  font-weight: 700;
-  font-size: 14px;
-  line-height: 22px;
-  display: flex;
-  align-items: center;
-  letter-spacing: -0.408px;
-  padding-left: 10px;
-  padding-right: 10px;
-  height: 25px;
-  background: #ffb45c;
-  border-radius: 20px;
-  color: #000000;
-}
