--- web/src/sangjin/components/MainPage/Trend.tsx
@@ -1,8 +1,0 @@
-import * as React from "react";
-export function Trend() {
-  return (
-    <div>
-      <div>나의 도박 기록</div>
-    </div>
-  );
-}
